inmates: x86: timing: use a wider type for delay_us
As pm_timer_read() + microsecs * NS_PER_USEC may overflow on 32 bit
systems. Use a 64 bit type.

+++ inmates/lib/x86/timing.c
@@ -107,9 +107,9 @@ unsigned long pm_timer_read(void)
 
 void delay_us(unsigned long microsecs)
 {
-	unsigned long timeout = pm_timer_read() + microsecs * NS_PER_USEC;
+	unsigned long long timeout = pm_timer_read() + microsecs * NS_PER_USEC;
 
-	while ((long)(timeout - pm_timer_read()) > 0)
+	while ((long long)(timeout - pm_timer_read()) > 0)
 		cpu_relax();
 }
